moved scss test to plugin-build-library-extended; cleaned up expected output.
