+++ packages/loom-plugin-build-library-extended/src/tests/fixtures/scss/first.scss
@@ -1,0 +1,5 @@
+$size: 10px;
+
+.First {
+  margin: $size;
+}
