+++ packages/loom-plugin-build-library-extended/src/tests/fixtures/scss/second.scss
@@ -1,0 +1,5 @@
+$size: 20px;
+
+.Second {
+  margin: $size;
+}
