+++ packages/loom-plugin-build-library-extended/src/tests/fixtures/scss/third.scss
@@ -1,0 +1,5 @@
+$size: 30px;
+
+.Third {
+  margin: $size;
+}
