--- packages/loom-plugin-build-library/src/tests/fixtures/scss/second.scss
@@ -1,19 +1,0 @@
-@import './helpers';
-
-.Second {
-  .Nav {
-    margin: rem(23.7px);
-    padding: 0;
-    list-style: none;
-  }
-
-  .Link {
-    display: inline-block;
-  }
-
-  .Table {
-    display: block;
-    padding: 6px 12px;
-    text-decoration: none;
-  }
-}
