--- packages/loom-plugin-build-library/src/tests/fixtures/scss/third.scss
@@ -1,19 +1,0 @@
-@import './helpers';
-
-.Third {
-  .Icon {
-    margin: rem(23.7px);
-    padding: 0;
-    list-style: none;
-  }
-
-  .Logo {
-    display: inline-block;
-  }
-
-  .Svg {
-    display: block;
-    padding: 6px 12px;
-    text-decoration: none;
-  }
-}
